requirements.txt
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
binance-futures-connector==3.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.1
hyperlink==21.0.0
idna==3.4
incremental==21.3.0
numpy==1.23.3
pandas==1.5.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.1.0
python-dateutil==2.8.2
pytz==2022.4
requests==2.28.1
service-identity==21.1.0
six==1.16.0
Twisted==22.8.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
urllib3==1.26.12
zope.interface==5.4.0
